@@ -31,24 +31,31 @@ dependencies:
   flutter:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  # State Management
+  provider: ^6.1.1
+  
+  # Backend
+  supabase_flutter: ^2.9.1
+  
+  # Environment
+  flutter_dotenv: ^5.1.0
+  
+  # UI Components
   cupertino_icons: ^1.0.8
-  provider: ^6.1.5
-  supabase_flutter: ^2.0.0
-  flutter_dotenv: ^5.1.0  # Per gestire le chiavi in modo sicuro
+  
+  # Utilities
   uuid: ^4.3.3
+  intl: ^0.19.0  # For number and date formatting
+  
+  # Network
+  connectivity_plus: ^5.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^5.0.0
+  
+  # Linting
+  flutter_lints: ^3.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -61,10 +68,9 @@ flutter:
   # the material Icons class.
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  # Assets that are part of the application
+  assets:
+    - .env
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
